+++ duzzy-editor/src/command/select.rs
@@ -1,4 +1,6 @@
-use crate::editor::Workspace;
+use std::borrow::Cow;
+
+use crate::{buffer::Buffer, editor::Workspace};
 
 pub(super) fn select_line(ws: &mut Workspace) {
     let buf = ws.cur_mut().buf_mut();
@@ -21,6 +23,25 @@ pub(super) fn select_line(ws: &mut Workspace) {
 
     buf.update_selection(end);
     buf.set_offset(ofs);
+}
+
+pub(super) fn selected_text(buf: &Buffer) -> Option<Cow<str>> {
+    let selection = buf.selection()?;
+    let (start, mut end) = selection.range();
+
+    if selection.head() > selection.anchor() {
+        let len_chars = buf.len_chars();
+        end = (end + 1).min(len_chars);
+    }
+
+    let slice = buf.text().slice(start..end);
+
+    let text = match slice.as_str() {
+        Some(s) => Cow::from(s),
+        None => Cow::from(slice.to_string()),
+    };
+
+    Some(text)
 }
 
 #[cfg(test)]
